refactoring

+++ src/mem.rs
@@ -98,7 +98,7 @@ impl GcHeap {
                         next: None,
 
                         #[cfg(debug_assertions)]
-                        dbg_scope_level: self.scope_max_depth() as u8,
+                        dbg_scope_level: self.scope_max_depth(),
                         #[cfg(debug_assertions)]
                         dbg_string: std::any::type_name::<T>().into(),
                     };
